chore: allow GitHub Actions version names to have underscores (#719)
* chore: allow multiple characters in GitHub Actions version matching

In the case that we have multiple characters in the `with` clause, in
case it's a version with a long name, we should handle this.

Note that the previous regex would only allow a single character i.e.

```yaml
with:
  x-version: ...
```

But would not handle

```yaml
with:
  something-version: ...
```

* chore: allow GitHub Actions version names to have underscores

Although they're not "idiomatic" style for GitHub Actions, which prefers
kebab-case, there are still cases where Actions will use snake_case, and
we should allow capturing these versions.